+++ src/screens/DetailsScreen.tsx
