+++ src/screens/SignupScreen.tsx
@@ -5,10 +5,11 @@ import {
   View
 } from 'react-native';
 import Toast from 'react-native-toast-message';
-import AuthFormComponent from '../../componets/AuthFormComponent';
-import { RootStackScreenProp } from '../../navigation/type';
-import { handleSignUpWithEmail } from '../../utils/FirebaseAuth';
+
 import styles from './style/styles';
+import { RootStackScreenProp } from '../navigation/type';
+import { handleSignUpWithEmail } from '../utils/FirebaseAuth';
+import AuthFormComponent from '../componets/AuthFormComponent';
 
 const SignupScreen = ({ navigation }: RootStackScreenProp<'Signup'>) => {
   const [username, setUsername] = useState('');
